feat: save unsuccessful events and retry later
that way we wont lose any events that were not successful as they are sent again when the user reloads the page.
